+++ genvalidity-bytestring/src/Data/GenValidity/ByteString.hs
@@ -16,24 +16,6 @@ import qualified Data.ByteString.Lazy as LB
 import qualified Data.ByteString.Lazy.Internal as LB
 import qualified Data.ByteString.Short as Short
 
-instance GenUnchecked SB.ByteString where
-    genUnchecked =
-        error $
-        unlines
-            [ "Data.GenValidity.ByteString.genUnchecked :: Strict.ByteString"
-            , "You probably do not want to use this."
-            , "You probably want to use 'genValid' instead."
-            , "See https://github.com/NorfairKing/validity/blob/master/docs/BYTESTRING.md"
-            ]
-    shrinkUnchecked =
-        error $
-        unlines
-            [ "Data.GenValidity.ByteString.shrinkUnchecked :: Strict.ByteString -> [Strict.ByteString]"
-            , "You probably do not want to use this."
-            , "You probably want to use 'shrinkValid' instead."
-            , "See https://github.com/NorfairKing/validity/blob/master/docs/BYTESTRING.md"
-            ]
-
 -- |
 --
 -- > genValid = SB.pack <$> genValid
@@ -64,28 +46,38 @@ shrinkTrulyUncheckedStrictByteString :: SB.ByteString -> [SB.ByteString]
 shrinkTrulyUncheckedStrictByteString (SB.PS p o l) =
     [SB.PS p o' l' | (o', l') <- shrinkUnchecked (o, l)]
 
-instance GenUnchecked LB.ByteString where
-    genUnchecked =
-        sized $ \n ->
-            case n of
-                0 -> pure LB.Empty
-                _ -> do
-                    (a, b) <- genSplit n
-                    sb <- resize a genUnchecked
-                    lb <- resize b genUnchecked
-                    pure $ LB.Chunk sb lb
-    shrinkUnchecked lb_ =
-        case lb_ of
-            LB.Empty -> []
-            (LB.Chunk sb lb) ->
-                LB.Empty :
-                [LB.Chunk sb' lb' | (sb', lb') <- shrinkUnchecked (sb, lb)]
-
 instance GenValid LB.ByteString where
     genValid = LB.pack <$> genValid
     shrinkValid = fmap LB.pack . shrinkValid . LB.unpack
 
-instance GenInvalid LB.ByteString
+-- | WARNING: Unchecked ByteStrings are __seriously__ broken.
+--
+-- See 'genTrulyUncheckedStrictByteString'
+genTrulyUncheckedLazyByteString :: Gen LB.ByteString
+genTrulyUncheckedLazyByteString =
+    sized $ \n ->
+        case n of
+            0 -> pure LB.Empty
+            _ -> do
+                (a, b) <- genSplit n
+                sb <- resize a genTrulyUncheckedStrictByteString
+                lb <- resize b genTrulyUncheckedLazyByteString
+                pure $ LB.Chunk sb lb
+
+shrinkTrulyUncheckedLazyByteString :: LB.ByteString -> [LB.ByteString]
+shrinkTrulyUncheckedLazyByteString lb_ =
+    case lb_ of
+        LB.Empty -> []
+        (LB.Chunk sb lb) ->
+            LB.Empty :
+            [ LB.Chunk sb' lb'
+            | (sb', lb') <-
+                  shrinkTuple
+                      shrinkTrulyUncheckedStrictByteString
+                      shrinkTrulyUncheckedLazyByteString
+                      (sb, lb)
+            ]
+
 
 instance GenUnchecked Short.ShortByteString where
     genUnchecked = Short.pack <$> genValid